Replace Int64 with Int

+++ src/utils.jl
@@ -331,7 +331,7 @@ function _show_prettytable(
     newline_at_end=false,
     kwargs...,
 )
-    alignment_anchor_regex = Dict{Int64,Vector{Regex}}(
+    alignment_anchor_regex = Dict{Int,Vector{Regex}}(
         i => [r"\.", r"e", r"^NaN$", r"Inf$"] for (i, (k, v)) in enumerate(pairs(data)) if
         (eltype(v) <: Real && !(eltype(v) <: Integer) && !_is_ess_label(k))
     )
